+++ Frends.AzureDataLake.DeleteFiles/Frends.AzureDataLake.DeleteFiles/Definitions/Options.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 
 namespace Frends.AzureDataLake.DeleteFiles.Definitions;
 
@@ -9,9 +10,18 @@ public class Options
 {
     /// <summary>
     /// True: Throw an exception.
-    /// False: Return Result with IsSucces=false and ErrorMessage with description of exception.
+    /// False: Return Result with Success=false and Error with description of exception.
     /// </summary>
     /// <example>true</example>
     [DefaultValue(true)]
-    public bool ThrowErrorOnFailure { get; init; } = true;
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Custom error message to use when ThrowErrorOnFailure is true or when returning an error result.
+    /// If empty, the original exception message is used.
+    /// </summary>
+    /// <example></example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
 }